[refact] Move sequencing logic to dedicated `Sequencer` class
It involves a bit of communication between the `Searcher` and the `Sequencer`, but I hope it makes it more maintainable.

A positive side effect is that the sequencing logic is now unit-testable. But this only applies to the “base” logic, not to the “load more” logic that is tied to pagination and thus still to the `Searcher`.